refactor: rename components and their groups
